feat(web): adapt storage api
- Add #reprobe
- Remove #refresh
- Move useful methods from proposal file.
